SOFA score label

The SOFA (Sequential Organ Failure Assessment) score is a commonly used assessment tool for tracking a patient's status during a stay at an ICU. Organ function is quantified by aggregating 6 individual scores, representing respiratory, cardiovascular, hepatic, coagulation, renal and neurological systems. The function sofa_score() is used as callback function to the sofa concept but is exported as there are a few arguments that can used to modify some aspects of the presented SOFA implementation. Internally, sofa_score() calls first sofa_window(), followed by sofa_compute() and arguments passed as ... will be forwarded to the respective internally called function.

sofa_score(..., interval = NULL)

sofa_window(
  tbl,
  pafi_wf = min_or_na,
  plt_wf = min_or_na,
  bili_wf = max_or_na,
  map_wf = min_or_na,
  dopa_wf = max_or_na,
  norepi_wf = max_or_na,
  dobu_wf = max_or_na,
  epi_wf = max_or_na,
  gcs_wf = min_or_na,
  crea_wf = max_or_na,
  urine_wf = min_or_na,
  explicit_wins = FALSE,
  worst_win_length = hours(24L)
)

sofa_compute(
  tbl,
  na_val = 0L,
  na_resp = na_val,
  na_plt = na_val,
  na_liver = na_val,
  na_cardio = na_val,
  na_cns = na_val,
  na_renal = na_val,
  impute_fun = NULL,
  keep_components = FALSE,
  by_ref = TRUE
)

Arguments

...
Data and further arguments passed to sofa_window() or sofa_compute()

interval
Time series interval (only used for checking consistency of input data, NULL will use the interval of the first data object)

tbl
Table holding SOFA covariates

pafi_wf, plt_wf, bili_wf, map_wf, dopa_wf, norepi_wf, dobu_wf, epi_wf, gcs_wf, crea_wf, urine_wf
functions used to calculate worst values over windows

explicit_wins
The default FALSE iterates over all time steps, TRUE uses only the last time step per patient and a vector of times will iterate over these explicit time points

worst_win_length
Time-frame to look back and apply the *_wfs

na_val
Value to use for missing data

na_resp, na_plt, na_liver, na_cardio, na_cns, na_renal
Feature-specific values to use in case of missing data; default is na_val

impute_fun
Function used to impute missing data; default is NULL

keep_components
Logical flag indicating whether to return individual SOFA components as columns

by_ref
Logical flag indicating whether to perform the operation by reference

Details

The function sofa_window() calculates, for each covariate, the worst value over a moving window as specified by worst_win_length, using the respective function passed as *_wf (e.g. pafi_wf for the pafi component. The default functions min_or_na() and max_or_na() return NA instead of -Inf/Inf in the case where no measurement is available over an entire window.

Using data imputed by sofa_window(), the function sofa_compute() performs the actual score calculation. First, for each time step and component, measurements are converted to a component score using the definition by Vincent et. al.:

SOFA score	1	2	3	4

Respiration

PaO2/FiO2 [mmHg]	< 400	< 300	< 200	< 100

and mechanical ventilation			yes	yes

Coagulation

Platelets [×103/mm3]	< 150	< 100	< 50	< 20

Liver

Bilirubin [mg/dl] ([μmol/μl])	1.2-1.9 (20-32)	2.0-5.9 (33-101)	6.0-11.9 (102-204)	> 12.0 (> 204)

Cardiovasculara

MAP	< 70 mmHg

or dopamine		≤ 5	> 5	> 15

or dobutamine		any dose

or epinephrine			≤ 0.1	> 0.1

or norepinephrine			≤ 0.1	> 0.1

Central nervous system

Glasgow Coma Score	13-14	10-12	6-9	< 6

Renal

Creatinine [mg/dl] ([μmol/μl])	1.2-1.9 (110-170)	2.0-3.4 (171-299)	3.5-4.9 (300-440)	> 5.0 (< 440)

or urine output [ml/day]			< 500	< 200

aAdrenergic agents administered for at least 1h (doses given are in [μg/kg · min]

In case, for a given time step and component, no measurement is available, the corresponding na_* value is used (e.g. na_resp for the respiratory component). At default, this is 0 (the lowest possible score for a SOFA component). It is possible to retain missingness by passing NA as na_* value and using a function passed as impute_fun() in order to perform an additional imputation step.

References

Vincent, J.-L., Moreno, R., Takala, J. et al. The SOFA (Sepsis-related Organ Failure Assessment) score to describe organ dysfunction/failure. Intensive Care Med 22, 707–710 (1996). https://doi.org/10.1007/BF01709751