Data attach utilities

Data attach utilities

attach_mimic(
  demo = FALSE,
  dir = mimic_data_dir(demo),
  config = mimic_config(demo),
  envir = if (demo) mimic_demo else mimic
)

attach_eicu(
  demo = FALSE,
  dir = eicu_data_dir(demo),
  config = eicu_config(demo),
  envir = if (demo) eicu_demo else eicu
)

attach_hirid(
  dir = data_dir("hirid", create = FALSE),
  config = get_config("hirid-setup")
)

data

mimic

mimic_demo

eicu

eicu_demo

hirid

attach_datasource(dir, config)

setup_proxy_env(config, envir = new.env())

Arguments

demo
Logical flag indicating whether to import the respective demo of full dataset. Only used for determining the correct default arguments passed on to attach_datasource().

dir
Data directory where the imported .fst files are located (see import_datasource() for further information).

config
Import configuration specified as nested list (see import_datasource() for further information).

envir
(Optional) environment that serves as proxy for the environment holding references to tables. Passing NULL foregoes setting up this forwarding (implemented as active bindings, see base::makeActiveBinding()).

Format

An object of class environment of length 5.