Data download utilities

The Laboratory for Computational Physiology (LCP) at MIT hosts two large-scale databases of hospital intensive care units (ICUs), both of which can be either downloaded in full or as demo subsets. While both databases are publicly available, full download requires credentialed access which can be gained by applying for up an account with PhysioNet .

download_source(x, dir = NULL, tables = NULL, ...)

Arguments

x
Object of which a get_src_config() method is defined

dir
Destination directory where the downloaded data is written to.

tables
Character vector specifying the tables to download. If NULL, all available tables are downloaded.

...
Passed onto keyring, for example keyring::key_set_with_value().

Details

The Medical Information Mart for Intensive Care (MIMIC) database holds detailed clinical data from roughly 60,000 patient stays in Beth Israel Deaconess Medical Center (BIDMC) intensive care units between 2001 and 2012. The database includes information such as demographics, vital sign measurements made at the bedside (~1 data point per hour), laboratory test results, procedures, medications, caregiver notes, imaging reports, and mortality (both in and out of hospital). For further information, please refer to the MIMIC-III documentation .

More recently, Philips Healthcare and LCP began assembling the eICU Collaborative Research Database as a multi-center resource for ICU data. Combining data of several critical care units throughout the continental United States from the years 2014 and 2015, this database contains deidentified health data associated with over 200,000 admissions, including vital sign measurements, care plan documentation, severity of illness measures, diagnosis information, and treatment information. For further information, please refer to the eICU documentation .

In case the keyring package is available, credentials used to log onto PhysioNet are stored using keyring::key_set(), therefore allowing for non- interactive querying of the PhysioNet service. Furthermore, if the openssl package is available, SHA256 hashes of downloaded files are verified using openssl::sha256().

References

MIMIC-III, a freely accessible critical care database. Johnson AEW, Pollard TJ, Shen L, Lehman L, Feng M, Ghassemi M, Moody B, Szolovits P, Celi LA, and Mark RG. Scientific Data (2016). DOI: 10.1038/sdata.2016.35.

The eICU Collaborative Research Database, a freely available multi-center database for critical care research. Pollard TJ, Johnson AEW, Raffa JD, Celi LA, Mark RG and Badawi O. Scientific Data (2018). DOI: http://dx.doi.org/10.1038/sdata.2018.178.

Examples

if (FALSE) {

dir <- tempdir()
list.files(dir)

download_datasource("mimic_demo", dir = dir)
list.files(dir)

unlink(dir, recursive = TRUE)

}