Internal utilities for item/concept objects

Several internal utilities for modifying, querying ans subsetting item and concept objects, including getters and setters for itm variables, callback functions, cncpt target classes, as well as utilities for data loading such as prepare_query() which creates a row-subsetting expression, do_callback(), which applies a callback function to data or do_itm_load(), which performs data loading corresponding to an itm

prepare_query(x)

try_add_vars(x, ..., var_lst = NULL, type = c("data_vars", "meta_vars"))

get_itm_var(x, var = NULL, type = c("data_vars", "meta_vars"))

set_callback(x, fun)

do_callback(x, ...)

do_itm_load(x, id_type = "icustay", interval = hours(1L))

n_tick(x)

set_target(x, target)

get_target(x)

subset_src(x, src)

# S3 method for item
subset_src(x, src)

# S3 method for cncpt
subset_src(x, src)

# S3 method for concept
subset_src(x, src)

Arguments

x

Object defining the row-subsetting

...

Variable specification

var_lst

List-based variable specification

type

Variable type (either data or meta)

var

Variable name (NULL) returns all available

fun

Callback function (passed as string)

id_type

String specifying the patient id type to return

interval

The time interval used to discretize time stamps with, specified as base::difftime() object

src

Character vector of data source name(s)

Value

prepare_query(): an unevaluated expression used for row-subsetting

try_add_vars(): a (potentially) modified item object with added variables

get_itm_var(): character vector of itm variables

set_callback(): a modified object with added callback function

do_callback(): result of the callback function applied to data, most likely (id_tbl/ts_tbl)

do_itm_load(): result of item loading (id_tbl/ts_tbl)

n_tick(): Integer valued number of progress bar ticks

set_target(): a modified object with newly set target class

get_target(): string valued target class of an object

subset_src(): an object of the same type as the object passed as x