Load concept data

Data specified as concept objects is loaded, aggregated per combination of id and time-step and merged into wide format.

load_concepts(concepts, aggregate = NA_character_, merge_data = TRUE, ...)

load_concept(concept, aggregate = NA_character_, na_rm = TRUE, ...)

Arguments

concepts
Vector of concept objects to be loaded

aggregate
Controls how data within concepts is aggregated

merge_data
Logical flag, specifying whether to merge concepts into wide format or return a list, each entry corresponding to a concept

...
Passed to load_concept() and on to load_items()

concept
Single concept object to be loaded

na_rm
Logical flag, indicating whether to remove NA rows