Sepsis 3 label

The sepsis 3 label consists of a suspected infection combined with an acute increase in SOFA score.

sepsis_3(
  sofa,
  si,
  si_window = c("first", "last", "any"),
  delta_fun = delta_cummin,
  sofa_thresh = 2L
)

delta_cummin(x)

delta_start(x)

delta_min(x, shifts = seq.int(0L, 23L))

Arguments

sofa
ts_tbl with a column sofa_score

si
ts_tbl with columns si_lwr and si_upr defining windows for which a suspected infection is valid

si_window
Switch that can be used to filter SI windows

delta_fun
Function used to determine the SOFA increase during an SI window

sofa_thresh
Required SOFA increase to trigger Sepsis 3

x
Vector of SOFA scores

shifts
Vector of time shifts (multiples of the current interval) over which base::pmin() is evaluated