Stays

Building on functionality offered by the (internal) function id_map(), stay windows as well as (in case of differing values being passed as id_type and win_type) an ID mapping is computed.

stay_windows(x, ...)

# S3 method for src_env
stay_windows(
  x,
  id_type = "icustay",
  win_type = id_type,
  in_time = "start",
  out_time = "end",
  interval = hours(1L),
  patient_ids = NULL,
  ...
)

# S3 method for character
stay_windows(x, ...)

# S3 method for list
stay_windows(x, ..., patient_ids = NULL)

# S3 method for default
stay_windows(x, ...)

Arguments

x
Data source (is coerced to src_env using as_src_env())

...
Generic consistency

id_type
Type of ID all returned times are relative to

win_type
Type of ID for which the in/out times is returned

in_time, out_time
column names of the returned in/out times

interval
The time interval used to discretize time stamps with, specified as base::difftime() object

patient_ids
Patient IDs used to subset the result

Value

An id_tbl containing the selected IDs and depending on values passed as in_time and out_time, start and end times of the ID passed as win_var.

See also