Table specification

Used when importing a data source, a table specification defines column data types, source file names, a mapping of source to imported column names and optionally a number of expected rows and a partitioning scheme.

new_table_spec(table, files, cols, nrow = NULL, partitioning = NULL)

is_table_spec(x)

as_table_spec(x)

# S3 method for table_spec
as_table_spec(x)

# S3 method for list
as_table_spec(x)

get_table_specs(x)

table_names(x)

# S3 method for src_config
table_names(x)

# S3 method for list
table_names(x)

# S3 method for table_spec
table_names(x)

file_names(x)

# S3 method for src_config
file_names(x)

# S3 method for list
file_names(x)

# S3 method for table_spec
file_names(x)

n_partitions(x)

# S3 method for src_config
n_partitions(x)

# S3 method for list
n_partitions(x)

# S3 method for table_spec
n_partitions(x)

fst_names(x)

get_col_spec(x)

check_n_rows(x, comparison)

get_col_names(x)

get_part_fun(x, orig_names = FALSE)

Arguments

table
String valued table name

cols
List containing a list per column each holding string valued entries name (column name as used by ricu), col (column name as used in the raw data) and spec (name of readr::cols() column specification). Further entries will be passed as argument to the respective readr column specification

nrow
A count indicating the expected number of rows

partitioning
A table partitioning is defined by a column name and a vector of numeric values that are passed as vec argument to base::findInterval()

x
Object to test/coerce

fies
A character vector of file names