ICU class data reshaping

Utilities for reshaping id_tbl and ts_tbl objects.

cbind_id_tbl(
  ...,
  keep.rownames = FALSE,
  check.names = FALSE,
  key = NULL,
  stringsAsFactors = FALSE
)

rbind_id_tbl(..., use.names = TRUE, fill = FALSE, idcol = NULL)

# S3 method for id_tbl
merge(x, y, by = NULL, by.x = NULL, by.y = NULL, ...)

# S3 method for id_tbl
split(x, ...)

rbind_lst(x, ...)

merge_lst(x)

unmerge(x, col_groups = as.list(data_vars(x)), by = meta_vars(x), na_rm = TRUE)

Arguments

...

Objects to combine

keep.rownames, check.names, key, stringsAsFactors

Forwarded to data.table::data.table

use.names, fill, idcol

Forwarded to data.table::rbindlist

x, y

Objects to combine

by, by.x, by.y

Column names used for combining data

col_groups

A list of character vectors defining the grouping of non-by columns

na_rm

Logical flag indicating whether to remove rows that have all missing entries in the respective col_groups group

Value

Either id_tbl or ts_tbl objects (depending on inputs) or lists thereof in case of split() and unmerge().