Methods for creating and inspecting ts_tbl objects

Methods for creating and inspecting ts_tbl objects

# S3 method for ts_tbl
[(x, ...)

# S3 method for ts_tbl
dimnames(x)

# S3 method for ts_tbl
print(x, ..., n = NULL, width = NULL, n_extra = NULL)

# S3 method for ts_tbl
format(x, ..., n = NULL, width = NULL, n_extra = NULL)

# S3 method for ts_tbl
tbl_sum(x)

# S3 method for ts_tbl
str(object, ...)

# S3 method for ts_tbl
as.data.table(x, ...)

# S3 method for ts_tbl
as.data.frame(x, row.names = NULL, optional = FALSE, ...)

ts_tbl(..., key, index = NULL, interval = hours(1L))

as_ts_tbl(tbl, key, index = NULL, interval = hours(1L))

Arguments

...
Passed to data.table::data.table()/generic compatibility.

object
A ts_tbl object.

row.names, optional
Generic consistency: passing anything other than the default value issues a warning.

key
Character or numeric vector of at least length 1, identifying columns that combined with time stamps, uniquely identify rows.

index
Character or numeric vector of length 1, identifying the column that defines temporal ordering.

interval
Scalar value, defining time steps between rows (assuming complete time series data).

tbl
An object inheriting from data.frame.