Data items

Item objects are used in ricu as a way to specify how individual data items corresponding to clinical concepts (see also concept()), such as heart rate can be loaded from a data source. Several functions are available for constructing item (and related auxillary) objects either from code or by parsing a JSON formatted concept dictionary using load_dictionary().

new_itm(src, ..., class = "sel_itm")

is_itm(x)

init_itm(x, ...)

# S3 method for sel_itm
init_itm(x, table, sub_var, ids, callback = "identity_callback", ...)

# S3 method for hrd_itm
init_itm(x, table, sub_var, ids, callback = "identity_callback", ...)

# S3 method for col_itm
init_itm(x, table, unit_val = NULL, callback = "identity_callback", ...)

# S3 method for rgx_itm
init_itm(x, table, sub_var, regex, callback = "identity_callback", ...)

# S3 method for fun_itm
init_itm(x, callback, ...)

# S3 method for itm
init_itm(x, ...)

new_item(x)

item(...)

as_item(x)

is_item(x)

Arguments

src
The data source name

...
Further specification of the itm object (passed to init_itm())

class
Sub class for customizing itm behavior

x
Object to query/dispatch on

table
Name of the table containing the data

sub_var
Column name used for subsetting

ids
Vector of ids used to subset table rows. If NULL, all rows are considered corresponding to the data item

callback
Name of a function to be called on the returned data used for data cleanup operations (or a string that evaluates to a function)

unit_val
String valued unit to be used in case no unit_var is available for the given table

regex
String-valued regular expression which will be evaluated by base::grepl() with ignore.case = TRUE

Details

In order to allow for a large degree of flexibility (and extensibility), which is much needed owing to considerable heterogeneity presented by different data sources, several nested S3 classes are involved in representing a concept. An outline of this hierarchy can be described as

concept: contains many cncpt objects (of potentially differing sub-types), each comprising of some meta-data and an item object

item: contains many itm objects (of potentially differing sub-types), each encoding how to retrieve a data item.

The design choice for wrapping a vector of itm objects with a container class item is motivated by the requirement of having several different sub-types of itm objects (all inheriting from the parent type itm), while retaining control over how this homogeneous w.r.t. parent type, but heterogeneous w.r.t. sub-type vector of objects behaves in terms of S3 generic functions.

The following sub-classes to itm are available, each representing a different data-scenario:

sel_itm: The most widely used item class is intended for the situation where rows of interest can be identified by looking for occurrences of a set of IDs (ids) in a column (sub_var). An example for this is heart rate hr on mimic, where the IDs 211 and 220045are looked up in theitemidcolumn ofchartevents`.

col_itm: This item class can be used if no row-subsetting is required. An example for this is heart rate (hr) on eicu, where the table vitalperiodic contains an entire column dedicated to heart rate measurements.

rgx_itm: As alternative to the value-matching approach of sel_itm objects, this class identifies rows using regular expressions. Used for example for insulin in eicu, where the regular expression ^insulin (250.+)?\\(((ml|units)/hr)?\\)$ is matched against the drugname column of infusiondrug. The regular expression is evaluated by base::grepl() with ignore.case = TRUE.

fun_itm: Intended for the scenario where data of interest is not directly available from a table, this itm class offers most flexbility. A function can be specified as callback and this function will be called with arguments x (the object itself), patient_ids, id_type and interval (see load_concepts()) and is expected to return an object as specified by the target entry.

All itm objects have to specify a data source (src) and a target type (target), which specifies the expected object

Examples

gluc <- item("mimic_demo", "labevents", "itemid", list(c(50809L, 50931L)))

is_item(gluc)

#> [1] TRUE

identical(gluc, as_item(load_dictionary("mimic_demo", "glu")))

#> [1] FALSE