File system utilities

Determine the location where to place data meant to persist between individual sessions.

data_dir(subdir = NULL, create = TRUE)

get_config(name, dir = NULL, ...)

set_config(x, name, dir = file.path("inst", "extdata", "config"), ...)

Arguments

subdir
A string specifying a directory that will be made sure to exist below the data directory.

create
Logical flag indicating whether to create the specified directory

name
File name of the configuration file (without file ending)

dir
(Optional) directory name where to look for the specified file

...
Passed to jsonlite::read_json() or jsonlite::write_json()

x
Object to be written

Details

For data, the default location depends on the operating system as

Platform	Location

Linux	~/.local/share/ricu

macOS	~/Library/Application Support/ricu

Windows	%LOCALAPPDATA%/ricu

If the default storage directory does not exists, it will only be created upon user consent (requiring an interactive session).

The environment variable RICU_DATA_PATH can be used to overwrite the default location. If desired, this variable can be set in an R startup file to make it apply to all R sessions. For example, it could be set within:

A project-local .Renviron;

The user-level .Renviron;

A file at $(R RHOME)/etc/Renviron.site.

Any directory specified as environment variable will recursively be created.

Examples

Sys.setenv(RICU_DATA_PATH = tempdir())
identical(data_dir(), tempdir())

#> [1] TRUE

dir.exists(file.path(tempdir(), "some_subdir"))

#> [1] FALSE

some_subdir <- data_dir("some_subdir")
dir.exists(some_subdir)

#> [1] TRUE

cfg <- get_config("concept-dict")
identical(
  cfg,
  get_config("concept-dict",
             system.file("extdata", "config", package = "ricu"))
)

#> [1] TRUE