Tabular ICU data classes

In order to simplify handling or tabular ICU data, ricu provides two S3 classes, id_tbl and ts_tbl. The two classes essentially consist of a data.table object, alongside some meta data and S3 dispatch is used to enable more natural behavior for some data manipulation tasks. For example, when merging two tables, a default for the by argument can be chosen more sensibly if columns representing patient ID and timestamp information can be identified.

id_tbl(..., id_vars = 1L)

is_id_tbl(x)

as_id_tbl(x, id_vars = NULL, by_ref = FALSE)

ts_tbl(..., id_vars = 1L, index_var = NULL, interval = NULL)

is_ts_tbl(x)

as_ts_tbl(x, id_vars = NULL, index_var = NULL, interval = NULL, by_ref = FALSE)

validate_tbl(x)

Arguments

...
forwarded to data.table::data.table() or generic consistency

id_vars
Column name(s) to be used as id column(s)

x
Object to query/operate on

by_ref
Logical flag indicating whether to perform the operation by reference

index_var
Column name of the index column

interval
Time series interval length specified as scalar-valued difftime object

Details

The two classes are designed for two often encountered data scenarios:

id_tbl objects can be used to represent static (with repspect to relevant time scales) patient data such as patient age and such an object is simply a data.table combined with a non-zero length character vector valued attribute marking the columns tracking patient ID information (id_vars). All further columns are considered as data_vars.

ts_tbl objects are used for grouped time series data. A data.table object again is augmented by attributes, including a non-zero length character vector identifying patient ID columns (id_vars), a string, tracking the column holding time-stamps (index_var) and a scalar difftime object determining the time-series step size interval. Again, all further columns are treated as data_vars.

Owing to the nested structure of required meta data, ts_tbl inherits from id_tbl. Furthermore, both classes inherit from data.table. As such, data.table reference semantics are available for some operations, indicated by presence of a by_ref argument. At default, value, by_ref is set to FALSE as this is in line with base R behavior at the cost of potentially incurring unnecessary data copies. Some care has to be taken when passing by_ref = TRUE and enabling by reference operations as this can have side effects (see examples).

For instantiating ts_tbl objects, both index_var and interval can be automatically determined if not specified. For the index column, the only requirement is that a single difftime column is present, while for the time step, the minimal difference between two consecutive observations is chosen (and all differences are therefore required to be multiples of the minimum difference).

Coercion between id_tbl and ts_tbl by default keeps intersecting attributes fixed and new attributes are by default inferred as for class instantiation. Each class comes with a class-specific implementation of the S3 generic function validate_tbl() which returns TRUE if the ibject is considered valid or a string outlining the type of validation failure that was encountered. Validity requires

inheriting from data.table and unique column names

for id_tbl that all columns specified by the non-zero length character vector holding onto the id_vars specification are available

for ts_tbl that the string-valued index_var column is available and does not intersect with id_vars and that the index column obeys the specified interval.

Finally, inheritance can be checked by calling is_id_tbl() and is_ts_tbl(). Note that due to ts_tbl inheriting from id_tbl, is_id_tbl() returns TRUE for both id_tbl and ts_tbl objects, while is_ts_tbl() only returns TRUE for ts_tbl objects.

Examples

tbl <- id_tbl(a = 1:10, b = rnorm(10))
is_id_tbl(tbl)

#> [1] TRUE

is_ts_tbl(tbl)

#> [1] FALSE

dat <- data.frame(a = 1:10, b = hours(1:10), c = rnorm(10))
tbl <- as_ts_tbl(dat, "a")

#> In order to automatically determine the index column, exactly one `difftime`
#> column is required.

is_id_tbl(tbl)

#> [1] TRUE

is_ts_tbl(tbl)

#> [1] TRUE

tmp <- as_id_tbl(tbl)
is_ts_tbl(tbl)

#> [1] TRUE

is_ts_tbl(tmp)

#> [1] FALSE

tmp <- as_id_tbl(tbl, by_ref = TRUE)
is_ts_tbl(tbl)

#> [1] FALSE

is_ts_tbl(tmp)

#> [1] FALSE