Tabular ICU data

In order to simplify handling or tabular ICU data, ricu provides two S3 classes, id_tbl and ts_tbl. The two classes essentially consist of a data.table object, alongside some meta data and S3 dispatch is used to enable more natural behavior for some data manipulation tasks. For example, when merging two tables, a default for the by argument can be chosen more sensibly if columns representing patient ID and timestamp information can be identified.

id_tbl(..., id_vars = 1L)

is_id_tbl(x)

as_id_tbl(x, id_vars = NULL, by_ref = FALSE)

# S3 method for ts_tbl
as_id_tbl(x, id_vars = NULL, by_ref = FALSE)

# S3 method for id_tbl
as_id_tbl(x, id_vars = NULL, by_ref = FALSE)

# S3 method for data.table
as_id_tbl(x, id_vars = NULL, by_ref = FALSE)

# S3 method for default
as_id_tbl(x, id_vars = NULL, by_ref = FALSE)

ts_tbl(..., id_vars = 1L, index_var = NULL, interval = NULL)

is_ts_tbl(x)

as_ts_tbl(x, id_vars = NULL, index_var = NULL, interval = NULL, by_ref = FALSE)

# S3 method for ts_tbl
as_ts_tbl(x, id_vars = NULL, index_var = NULL, interval = NULL, by_ref = FALSE)

# S3 method for id_tbl
as_ts_tbl(x, id_vars = NULL, index_var = NULL, interval = NULL, by_ref = FALSE)

# S3 method for data.table
as_ts_tbl(x, id_vars = NULL, index_var = NULL, interval = NULL, by_ref = FALSE)

# S3 method for default
as_ts_tbl(x, id_vars = NULL, index_var = NULL, interval = NULL, by_ref = FALSE)

unclass_tbl(x)

# S3 method for data.frame
unclass_tbl(x)

# S3 method for ts_tbl
unclass_tbl(x)

# S3 method for id_tbl
unclass_tbl(x)

reclass_tbl(x, template, stop_on_fail = TRUE)

# S3 method for `NULL`
reclass_tbl(x, template, stop_on_fail = TRUE)

# S3 method for id_tbl
reclass_tbl(x, template, stop_on_fail = TRUE)

# S3 method for ts_tbl
reclass_tbl(x, template, stop_on_fail = TRUE)

as_ptype(x)

# S3 method for id_tbl
as_ptype(x)

# S3 method for ts_tbl
as_ptype(x)

validate_tbl(x)

# S3 method for id_tbl
validate_tbl(x)

# S3 method for ts_tbl
validate_tbl(x)

# S3 method for data.table
validate_tbl(x)

Arguments

...
forwarded to data.table::data.table() or generic consistency

id_vars
Column name(s) to be used as id column(s)

x
Object to query/operate on

by_ref
Logical flag indicating whether to perform the operation by reference

index_var
Column name of the index column

interval
Time series interval length specified as scalar-valued difftime object

template
Object after which to model the object in question

stop_on_fail
Logical flag indicating whether to consider failed object validation as error

Details

The two classes are designed for two often encoutered data scenatios:

id_tbl objects can be used to represent static (with repspect to relevant time scales) patient data such as patient age and such an object is simply a data.table combined with a non-zero length character vector valued attribute marking the columns tracking patient ID information (id_vars). All further columns are considered as data_vars.

ts_tbl objects are used for grouped time series data. A data.table object again is augmented by attributes, including a non-zero length character vector identifying patient ID columns (id_vars), a string, tracking the column holding time-stamps (index_var) and a scalar difftime object determining the time-series step size interval. Again, all further columns are treated as data_vars.

Owing to the nested structure of required meta data, ts_tbl inherits from ìd_tbl. Furthermore, both classes inherit from data.table. As such, data.table[reference semantics][data.table::reference] are available for some operations, indicated by presence of aby_refargument. At default, value,by_refis disabled as this is inline with base R behavior at the cost of potentially incurring unnecessary data copies. Some care has to be taken when passingby_ref = TRUE` and enabling by reference operations as this can have side effects (see examples).