Suspicion of infection label

Suspected infection is defined as co-occurrence of of antibiotic treatment and body-fluid sampling.

susp_inf(
  ...,
  abx_count_win = hours(24L),
  abx_min_count = 1L,
  positive_cultures = FALSE,
  si_mode = c("and", "or", "abx", "samp"),
  abx_win = hours(24L),
  samp_win = hours(72L),
  by_ref = TRUE,
  keep_components = FALSE,
  interval = NULL
)

Arguments

...
Data and further arguments are passed to si_calc()

abx_count_win
Time span during which to apply the abx_min_count criterion

abx_min_count
Minimal number of antibiotic administrations

positive_cultures
Logical flag indicating whether to require cultures to be positive

si_mode
Switch between and, or, abx, samp modes

abx_win
Time-span within which sampling has to occur

samp_win
Time-span within which antibiotic administration has to occur

by_ref
Logical flag indicating whether to process data by reference

keep_components
Logical flag indicating whether to return the individual components alongside the aggregated score

interval
Time series interval (only used for checking consistency of input data)

Details

Suspected infection can occur in one of the two following ways:

administration of antibiotics followed by a culture sampling within samp_win hours

       abx_win
   |---------------|
  ABX           sampling (last possible)

culture sampling followed by an antibiotic administration within abx_win hours

                     samp_win
   |---------------------------------------------|
sampling                                        ABX (last possible)

The default values of samp_win and abx_win are 24 and 72 hours respectively, as per Singer et.al. .

The earlier of the two times (fluid sampling, antibiotic treatment) is taken as the time of suspected infection (SI time). The suspected infection window (SI window) is defined to start si_lwr hours before the SI time and end si_upr hours after the SI time. The default values of 48 and 24 hours (respectively) are chosen as used by Seymour et.al. (see Supplemental Material).

                48h                       24h
  |------------------------------(|)---------------|
                                SI time

For some datasets, however, information on body fluid sampling is not available for majority of the patients (eICU data). Therefore, an alternative definition of suspected infection is required. For this, we use administration of multiple antibiotics (argument abx_min_count determines the required number) within abx_count_win hours. The first time of antibiotic administration is taken as the SI time in this case.

References

Singer M, Deutschman CS, Seymour CW, et al. The Third International Consensus Definitions for Sepsis and Septic Shock (Sepsis-3). JAMA. 2016;315(8):801–810. doi:10.1001/jama.2016.0287

Seymour CW, Liu VX, Iwashyna TJ, et al. Assessment of Clinical Criteria for Sepsis: For the Third International Consensus Definitions for Sepsis and Septic Shock (Sepsis-3). JAMA. 2016;315(8):762–774. doi:10.1001/jama.2016.0288