Suspicion of infection label

The Sepsis-3 label is a combination of a suspected infection and an acute increase in SOFA score by ≥ 2 points.

si_data(
  source,
  abx_count_win = hours(24L),
  abx_min_count = 1L,
  positive_cultures = FALSE,
  dictionary = read_dictionary("concept-dict"),
  ...
)

si_windows(
  tbl,
  si_mode = c("and", "or"),
  abx_win = hours(24L),
  samp_win = hours(72L),
  si_lwr = hours(48L),
  si_upr = hours(24L)
)

si(source, ...)

Arguments

source
String valued name of data source

abx_count_win
Time span during which to apply the abx_min_count criterion

abx_min_count
Minimal number of antibiotic administrations

positive_cultures
Logical flag indicating whether to require cultures to be positive

dictionary
dictionary object to use for concept lookup

...
Passed to load_concepts()

tbl
ts_tbl object to use for label computations

si_mode
Switch between and/or modes

abx_win
Time-span within which sampling has to occur

samp_win
Time-span within which antibiotic administration has to occur

si_lwr, si_upr
Lower/upper extent of SI windows