Low level functions for loading data

Data loading involves a cascade of S3 generic functions, which can individually be adapted to the specifics of individual data sources. A the lowest level, load_scr is called, followed by load_difftime(). Functions up the chain, are described in load_id().

load_src(x, ...)

# S3 method for src_tbl
load_src(x, rows, cols = colnames(x), ...)

# S3 method for character
load_src(x, src, ...)

load_difftime(x, ...)

# S3 method for mimic_tbl
load_difftime(
  x,
  rows,
  cols = colnames(x),
  id_hint = id_vars(x),
  time_vars = ricu::time_vars(x),
  ...
)

# S3 method for eicu_tbl
load_difftime(
  x,
  rows,
  cols = colnames(x),
  id_hint = id_vars(x),
  time_vars = ricu::time_vars(x),
  ...
)

# S3 method for hirid_tbl
load_difftime(
  x,
  rows,
  cols = colnames(x),
  id_hint = id_vars(x),
  time_vars = ricu::time_vars(x),
  ...
)

# S3 method for character
load_difftime(x, src, ...)

Arguments

x
Object for which to load data

...
Generic consistency

rows
Expression used for row subsetting (NSE)

cols
Character vector of column names

src
Passed to as_src_tbl() in order to determine the data source

id_hint
String valued id column selection (not necessarily honored)

time_vars
Character vector enumerating the columns to be treated as timestamps and thus returned as base::difftime() vectors

Details

A function extending the S3 generic load_src() is expected to load a subset of rows/columns from a tabular data source. While the column specification is provided as character vector of column names, the row subsetting involves non-standard evaluation (NSE). Data-sets that are included with ricu are represented by prt objects, which use rlang::eval_tidy() to evaluate NSE expressions. Furthermore, prt objects potentially represent tabular data split into partitions and row-subsetting expressions are evaluated per partition (see the part_safe flag in prt::subset.prt()). The return value of load_src() is expected to be of type data.table.

Timestamps are represented differently among the included data sources: while MIMIC-III and HiRID use absolute date/times, eICU provides temporal information as minutes relative to ICU admission. Other data sources, such as the ICU dataset provided by Amsterdam UMC, opt for relative times as well, but not in minutes since admission, but in milliseconds. In order to smoothen out such discrepancies, the next function in the data loading hierarchy is load_difftime(). This function is expected to call load_src() in order to load a subset of rows/columns from a table stored on disk and convert all columns that represent timestamps (as specified by the argument time_vars) into base::difftime() vectors using mins as time unit.

The returned object should be of type id_tbl, with the ID vars identifying the ID system the times are relative to. If for example all times are relative to ICU admission, the ICU stay ID should be returned as ID column. The argument id_hint may suggest an ID type, but if in the raw data, this ID is not available, load_difftime() may return data using a different ID system. In MIMIC-III, for example, data in the labevents table is available for subject_id (patient ID) pr hadm_id (hospital admission ID). If data is requested for icustay_id (ICU stay ID), this request cannot be fulfilled and data is returned using the ID system with the highest cardinality (among the available ones). Utilities such as change_id() can the later be used to resolve data to icustay_id.

Examples

tbl <- mimic_demo$labevents

#> Cannot update read-only data source `mimic_demo`

col <- c("charttime", "value")

load_src(tbl, itemid == 50809)

#>        row_id subject_id hadm_id itemid           charttime value valuenum
#>   1:  6245910      10006      NA  50809 2165-02-22 11:36:00    77       77
#>   2:  6250050      10019  177759  50809 2163-05-14 20:07:00   146      146
#>   3:  6250095      10019  177759  50809 2163-05-14 22:12:00   158      158
#>   4:  6250102      10019  177759  50809 2163-05-14 23:11:00   150      150
#>   5:  6250113      10019  177759  50809 2163-05-15 00:00:00   159      159
#>  ---                                                                      
#> 279: 20434896      44083  125157  50809 2112-05-04 17:07:00   144      144
#> 280: 20434911      44083  125157  50809 2112-05-04 21:05:00   163      163
#> 281: 20439454      44154  174245  50809 2178-05-14 16:52:00   164      164
#> 282: 20448342      44212  163189  50809 2123-12-09 12:25:00   117      117
#> 283: 20448362      44212  163189  50809 2123-12-09 18:09:00    99       99
#>      valueuom     flag
#>   1:    mg/dL     <NA>
#>   2:    mg/dL abnormal
#>   3:    mg/dL abnormal
#>   4:    mg/dL abnormal
#>   5:    mg/dL abnormal
#>  ---                  
#> 279:    mg/dL abnormal
#> 280:    mg/dL abnormal
#> 281:    mg/dL abnormal
#> 282:    mg/dL abnormal
#> 283:    mg/dL     <NA>

colnames(
  load_src("labevents", "mimic_demo", itemid == 50809, cols = col)
)

#> [1] "charttime" "value"    

load_difftime(tbl, itemid == 50809)

#> No overlap between configured id var options and available columns for table
#> `labevents` in mimic_demo

#> # An `id_tbl`: 276 ✖ 9
#> # Id var:      `hadm_id`
#>     hadm_id   row_id subject_id itemid charttime value valuenum valueuom flag   
#>       <int>    <int>      <int>  <int> <drtn>    <chr>    <dbl> <chr>    <chr>  
#> 1    100969 20237172      42430  50809 -177 mins 273        273 mg/dL    abnorm…
#> 2    104697 19993476      40612  50809  128 mins 118        118 mg/dL    abnorm…
#> 3    107689 20173434      41983  50809 -178 mins 105        105 mg/dL    NA     
#> 4    110244  6253811      10035  50809 1932 mins 105        105 mg/dL    NA     
#> 5    110958 20416364      43927  50809 -220 mins 99          99 mg/dL    NA     
#> …                    
#> 272  199395  6251511      10027  50809 5488 mins 152        152 mg/dL    abnorm…
#> 273  199395  6251524      10027  50809 6705 mins 146        146 mg/dL    abnorm…
#> 274  199395  6251546      10027  50809 6960 mins 108        108 mg/dL    abnorm…
#> 275  199395  6251577      10027  50809 8270 mins 58          58 mg/dL    abnorm…
#> 276  199395  6251586      10027  50809 8538 mins 88          88 mg/dL    NA     
#> # … with 266 more rows

colnames(
  load_difftime(tbl, itemid == 50809, col)
)

#> No overlap between configured id var options and available columns for table
#> `labevents` in mimic_demo

#> [1] "hadm_id"   "charttime" "value"    

id_vars(
  load_difftime(tbl, itemid == 50809, id_hint = "icustay_id")
)

#> No overlap between configured id var options and available columns for table
#> `labevents` in mimic_demo

#> [1] "hadm_id"

id_vars(
  load_difftime(tbl, itemid == 50809, id_hint = "subject_id")
)

#> [1] "subject_id"