Read and write utilities

Support for reading from and writing to .psv files as used for the Physionet Sepsis Challenge.

Tabular ICU data as used by ricu frequently contains a column that defines a grouping structure (e.g. patient IDs, ICU stay IDs, etc.) and a column holding time information.

write_psv(x, dir, na_rows = NULL)

read_psv(dir, col_spec = NULL, id_col = "stay_id")

new_id_meta(tbl_id)

is_id_meta(x)

as_id_meta(x)

# S3 method for id_meta
as_id_meta(x)

# S3 method for ts_meta
as_id_meta(x)

# S3 method for id_meta
meta_cols(x)

# S3 method for id_meta
tbl_class(x)

# S3 method for id_meta
validate_meta(x, meta)

new_tbl_index(index, interval)

new_tbl_id(id, opts = NULL)

is_tbl_index(x)

is_tbl_id(x)

# S3 method for tbl_index
tbl_index(x)

# S3 method for tbl_id
tbl_id(x)

new_ts_meta(tbl_id, tbl_index)

is_ts_meta(x)

# S3 method for ts_meta
validate_meta(x, meta)

is_tbl_meta(x)

tbl_meta(x)

# S3 method for tbl_meta
tbl_meta(x)

# S3 method for tbl_meta
tbl_id(x)

validate_meta(x, meta)

# S3 method for icu_tbl
tbl_meta(x)

# S3 method for icu_tbl
tbl_id(x)

tbl_index(x)

tbl_id(x)

tbl_class(x)

# S3 method for icu_tbl
tbl_class(x)

Arguments

x
Object to query/coerce

dir
Directory to write the (many) files to or read from

na_rows
If TRUE missing time steps are filled with NaN values, if FALSE, rows where all data columns entries are missing are removed and if NULL, data is written as-is

col_spec
A column specification as created by readr::cols()

id_col
Name of the id column (IDs are generated from file names)

tbl_id, tbl_index
Meta data objects to construct tbl_meta objects with

meta
Meta data object to validate

index
Column name of the index column

interval
Minimal time interval between rows

id
Column name of the ID column

opts
Alternative ID systems

References

Reyna, M., Josef, C., Jeter, R., Shashikumar, S., Moody, B., Westover, M. B., Sharma, A., Nemati, S., & Clifford, G. (2019). Early Prediction of Sepsis from Clinical Data -- the PhysioNet Computing in Cardiology Challenge 2019 (version 1.0.0). PhysioNet. https://doi.org/10.13026/v64v-d857.