Utilities for working with id_tbl and ts_tbl objects

Utilities for working with id_tbl and ts_tbl objects

id_vars(x)

# S3 method for id_tbl
id_vars(x)

id_var(x)

id_col(x)

index_var(x)

# S3 method for ts_tbl
index_var(x)

index_col(x)

meta_vars(x)

# S3 method for id_tbl
meta_vars(x)

# S3 method for ts_tbl
meta_vars(x)

data_vars(x)

data_var(x)

data_col(x)

interval(x)

# S3 method for ts_tbl
interval(x)

# S3 method for difftime
interval(x)

time_unit(x)

time_step(x)

time_vars(x)

# S3 method for data.table
time_vars(x)

rename_cols(x, new, old = colnames(x), skip_absent = FALSE, by_ref = FALSE)

# S3 method for ts_tbl
rename_cols(x, new, old = colnames(x), skip_absent = FALSE, ...)

# S3 method for id_tbl
rename_cols(x, new, old = colnames(x), skip_absent = FALSE, ...)

# S3 method for data.table
rename_cols(x, new, old = colnames(x), skip_absent = FALSE, by_ref = FALSE)

rm_cols(x, cols, skip_absent = FALSE, by_ref = FALSE)

change_interval(x, new_interval, cols = time_vars(x), by_ref = FALSE)

# S3 method for ts_tbl
change_interval(x, new_interval, cols = time_vars(x), ...)

# S3 method for data.table
change_interval(x, new_interval, cols = time_vars(x), by_ref = FALSE)

rbind_lst(x, ...)

rm_na(x, cols = data_vars(x), mode = c("all", "any"))

unmerge(x, col_groups = as.list(data_vars(x)), by = meta_vars(x), na_rm = TRUE)

dt_gforce(
  x,
  fun = c("mean", "median", "min", "max", "sum", "prod", "var", "sd", "first", "last"),
  by = meta_vars(x),
  vars = data_vars(x),
  na_rm = !fun %in% c("first", "last")
)

is_unique(x, ...)

# S3 method for default
is_unique(x, ...)

make_unique(
  x,
  expr = NULL,
  by = meta_vars(x),
  vars = data_vars(x),
  env = NULL,
  ...
)

Arguments

x
Object to query/operate on

new, old
Replacement names and existing column names for renaming columns

skip_absent
Logical flag for ignoring non-existent column names

by_ref
Logical flag indicating whether to perform the operation by reference

...
Ignored

cols
Column names of columns to consider

new_interval
Replacement interval length specified as scalar-valued difftime object

mode
Switch between all where all entries of a row have to be missing (for the selected columns) or any, where a single missing entry suffices

col_groups
A list of character vectors defining the grouping of non-by columns

by
Columns that will be present in every one of the resulting tables

na_rm
Logical flag indicating whether to remove rows that have all missing entries in the respective col_groups group

fun
Function name (as string) to apply over groups

vars
Column names to apply the function to

expr
Expression to apply over groups

env
Environment to look up names in expr