Utilities for icu_tbl objects

Utilities for icu_tbl objects

# S3 method for id_meta
rename_cols(x, new, old, ...)

# S3 method for tbl_index
rename_cols(x, new, old, ...)

# S3 method for tbl_id
rename_cols(x, new, old, ...)

# S3 method for ts_meta
rename_cols(x, new, old, ...)

rm_cols(x, cols, ...)

# S3 method for data.table
rm_cols(x, cols, ...)

rename_cols(x, new, old = colnames(x), ...)

# S3 method for data.table
rename_cols(x, new, old = colnames(x), ...)

# S3 method for icu_tbl
rm_cols(x, cols, ...)

# S3 method for icu_tbl
rename_cols(x, new, old = colnames(x), skip_absent = FALSE, ...)

rbind_lst(x, ...)

rm_na(x, cols = data_cols(x), mode = c("all", "any"))

unmerge(x, col_groups = as.list(data_cols(x)), by = meta_cols(x), na_rm = TRUE)

dt_gforce(
  x,
  fun = c("mean", "median", "min", "max", "sum", "prod", "var", "sd", "first", "last"),
  by = meta_cols(x),
  cols = data_cols(x),
  na_rm = !fun %in% c("first", "last")
)

is_unique(x, ...)

# S3 method for default
is_unique(x, ...)

make_unique(x, expr, fun, ...)

make_unique_quo(x, expr, by = meta_cols(x), cols = data_cols(x), ...)

Arguments

x
Object to operate on

new
New column names

old
Column names to be replaces by the ones specified as new

...
Generic consistency

cols
Column names of columns to consider

skip_absent
Silently ignore absent columns in old (along with their corresponding replacements in new)

mode
Switch between all where all entries of a row have to be missing (for the selected columns) or any, where a single missing entry suffices

col_groups
A list of character vectors defining the grouping of non-by columns

by
Columns that will be present in every one of the resulting tables

na_rm
Logical flag indicating whether to remove rows that have all missing entries in the respective col_groups group

fun
Function name (as string) to apply over groups

expr
Expression to apply over groups