Time series utility functions

Time series utility functions

expand_limits(
  x,
  min_col = "min",
  max_col = "max",
  step_size = 1L,
  id_cols = NULL,
  new_col = "hadm_time"
)

has_no_gaps(x)

has_gaps(...)

fill_gaps(x, limits = NULL, ...)

slide(x, expr, ...)

slide_quo(x, expr, before, after = hours(0L), ...)

slide_index(x, expr, ...)

slide_index_quo(x, expr, index, before, after = hours(0L), ...)

hop(x, expr, ...)

hop_quo(
  x,
  expr,
  windows,
  full_window = FALSE,
  lwr_col = "min_time",
  upr_col = "max_time"
)

Arguments

x
ts_tbl object to use

min_col, max_col
Name of the columns that represent lower and upper windows bounds

step_size
Controls the step size used to interpolate between min_col and max_col

id_cols
Names of the id columns

new_col
Name of the new index column

...
Passed to the *_quo version and from there to hop_quo()

limits
A table with columns for lower and upper window bounds

expr
Expression (quoted for *_quo and unquoted otherwise) to be evaluated over each window

before, after
Time span to look back/forward

index
A vector of times around which windows are spanned (relative to the index)

windows
An icu_tbl defining the windows to span

full_window
Logical flag controlling how the situation is handled where the sliding window extends beyond available data

lwr_col, upr_col
Names of columns (in windows) of lower/upper window bounds